HCP_fmripredict/requirements_update.txt
# packages in environment at /home/yuzhang/miniconda3/envs/tensorflow:
#
# Name                    Version                   Build  Channel
awscli                    1.14.47                   <pip>
bleach                    1.5.0                     <pip>
ca-certificates           2018.03.07                    0  
certifi                   2018.1.18                py36_0  
ciftify                   1.0.1                     <pip>
colorama                  0.3.7                     <pip>
cycler                    0.10.0                    <pip>
decorator                 4.2.1                     <pip>
docopt                    0.6.2                     <pip>
docutils                  0.14                      <pip>
entrypoints               0.2.3                     <pip>
h5py                      2.7.1                     <pip>
html5lib                  0.9999999                 <pip>
ipykernel                 4.8.1                     <pip>
ipython                   6.2.1                     <pip>
ipython-genutils          0.2.0                     <pip>
ipywidgets                7.1.1                     <pip>
jedi                      0.11.1                    <pip>
Jinja2                    2.10                      <pip>
jmespath                  0.9.3                     <pip>
jsonschema                2.6.0                     <pip>
jupyter                   1.0.0                     <pip>
jupyter-client            5.2.2                     <pip>
jupyter-console           5.2.0                     <pip>
jupyter-core              4.4.0                     <pip>
jupyterlab                0.31.8                    <pip>
jupyterlab-launcher       0.10.5                    <pip>
jupyterlmod               1.5.0                     <pip>
Keras                     2.1.4                     <pip>
libedit                   3.1                  heed3624_0  
libffi                    3.2.1                hd88cf55_4  
libgcc-ng                 7.2.0                hdf63c60_3  
libstdcxx-ng              7.2.0                hdf63c60_3  
lmdb                      0.93                      <pip>
Markdown                  2.6.11                    <pip>
MarkupSafe                1.0                       <pip>
matplotlib                2.1.2                     <pip>
mistune                   0.8.3                     <pip>
msgpack                   0.5.6                     <pip>
msgpack-numpy             0.4.3                     <pip>
nbconvert                 5.3.1                     <pip>
nbformat                  4.4.0                     <pip>
nbrsessionproxy           0.6.1                     <pip>
nbserverproxy             0.5.1                     <pip>
ncurses                   6.0                  h9df7e31_2  
nibabel                   2.2.1                     <pip>
nilearn                   0.4.0                     <pip>
notebook                  5.4.0                     <pip>
numexpr                   2.6.4                     <pip>
openssl                   1.0.2o               h20670df_0  
pandas                    0.21.0                    <pip>
pandocfilters             1.4.2                     <pip>
parso                     0.1.1                     <pip>
pexpect                   4.4.0                     <pip>
pickleshare               0.7.4                     <pip>
pip                       9.0.3                    py36_0  
prompt-toolkit            1.0.15                    <pip>
ptyprocess                0.5.2                     <pip>
pyasn1                    0.4.2                     <pip>
Pygments                  2.2.0                     <pip>
pyparsing                 2.2.0                     <pip>
python                    3.6.5                hc3d631a_0  
python-dateutil           2.6.1                     <pip>
pytz                      2018.3                    <pip>
PyYAML                    3.12                      <pip>
pyzmq                     17.0.0                    <pip>
qtconsole                 4.3.1                     <pip>
readline                  7.0                  ha6073c6_4  
rsa                       3.4.2                     <pip>
s3transfer                0.1.13                    <pip>
scikit-build              0.6.1                     <pip>
scikit-learn              0.19.1                    <pip>
scipy                     1.0.0                     <pip>
seaborn                   0.8.1                     <pip>
Send2Trash                1.4.2                     <pip>
setuptools                39.0.1                   py36_0  
simplegeneric             0.8.1                     <pip>
six                       1.11.0                    <pip>
sklearn                   0.0                       <pip>
sqlite                    3.22.0               h1bed415_0  
tables                    3.4.2                     <pip>
tabulate                  0.8.2                     <pip>
tensorboard               1.6.0                     <pip>
tensorflow-tensorboard    1.5.1                     <pip>
tensorpack                0.8.2                     <pip>
termcolor                 1.1.0                     <pip>
terminado                 0.8.1                     <pip>
testpath                  0.3.1                     <pip>
tk                        8.6.7                hc745277_3  
tornado                   4.5.3                     <pip>
tqdm                      4.19.7                    <pip>
traitlets                 4.3.2                     <pip>
wcwidth                   0.1.7                     <pip>
webencodings              0.5.1                     <pip>
Werkzeug                  0.14.1                    <pip>
wheel                     0.31.0                   py36_0  
widgetsnbextension        3.1.3                     <pip>
xz                        5.2.3                h55aa19d_2  
zlib                      1.2.11               ha838bed_2  
